# backend/docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"Product": {
				"properties": {
					"name": {
						"type": "string"
					},
					"price": {
						"type": "string"
					},
					"photos": {
						"items": {
							"$ref": "#/components/schemas/Photo"
						},
						"type": "array"
					},
					"description": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"country": {
						"type": "string"
					},
					"color": {
						"type": "string"
					},
					"brand": {
						"type": "string"
					},
					"orderItems": {
						"items": {
							"$ref": "#/components/schemas/Order"
						},
						"type": "array"
					},
					"visibility": {
						"type": "boolean"
					}
				},
				"required": [
					"photos",
					"country",
					"color",
					"brand",
					"orderItems",
					"visibility"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Photo": {
				"properties": {
					"url": {
						"type": "string"
					},
					"product": {
						"$ref": "#/components/schemas/Product"
					}
				},
				"required": [
					"url",
					"product"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Item": {
				"properties": {
					"productId": {
						"type": "number",
						"format": "double"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"product": {
						"$ref": "#/components/schemas/Product"
					}
				},
				"required": [
					"productId",
					"quantity",
					"product"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Order": {
				"properties": {
					"customerName": {
						"type": "string"
					},
					"customerEmail": {
						"type": "string"
					},
					"customerPhone": {
						"type": "string"
					},
					"customerAddress": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"orderItems": {
						"items": {
							"$ref": "#/components/schemas/Item"
						},
						"type": "array"
					}
				},
				"required": [
					"customerName",
					"customerEmail",
					"customerPhone",
					"customerAddress",
					"status",
					"orderItems"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "SPO_Practice-1",
		"version": "1.0.0",
		"description": "lfdjlgshif'kllk",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/order/{id}/updateStatus": {
			"put": {
				"operationId": "UpdateOrderStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"updatedOrder": {},
												"status": {
													"type": "number",
													"format": "double"
												},
												"error": {
													"type": "string"
												},
												"success": {
													"type": "boolean"
												}
											},
											"required": [
												"status",
												"error",
												"success"
											],
											"type": "object"
										},
										{
											"properties": {
												"status": {},
												"error": {},
												"updatedOrder": {
													"properties": {
														"status": {
															"type": "string"
														},
														"customerAddress": {
															"type": "string"
														},
														"customerPhone": {
															"type": "string"
														},
														"customerEmail": {
															"type": "string"
														},
														"customerName": {
															"type": "string"
														},
														"orderDate": {
															"type": "string",
															"format": "date-time"
														},
														"id": {
															"type": "number",
															"format": "double"
														}
													},
													"required": [
														"status",
														"customerAddress",
														"customerPhone",
														"customerEmail",
														"customerName",
														"orderDate",
														"id"
													],
													"type": "object"
												},
												"success": {
													"type": "boolean"
												}
											},
											"required": [
												"updatedOrder",
												"success"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Orders"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"newStatus": {
										"type": "string"
									}
								},
								"required": [
									"newStatus"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/order/add": {
			"post": {
				"operationId": "CreateOrder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"status": {},
												"error": {},
												"newOrder": {
													"allOf": [
														{
															"properties": {
																"orderItems": {
																	"items": {
																		"properties": {
																			"orderId": {
																				"type": "number",
																				"format": "double"
																			},
																			"productId": {
																				"type": "number",
																				"format": "double"
																			},
																			"quantity": {
																				"type": "number",
																				"format": "double"
																			},
																			"id": {
																				"type": "number",
																				"format": "double"
																			}
																		},
																		"required": [
																			"orderId",
																			"productId",
																			"quantity",
																			"id"
																		],
																		"type": "object"
																	},
																	"type": "array"
																}
															},
															"required": [
																"orderItems"
															],
															"type": "object"
														},
														{
															"properties": {
																"status": {
																	"type": "string"
																},
																"customerAddress": {
																	"type": "string"
																},
																"customerPhone": {
																	"type": "string"
																},
																"customerEmail": {
																	"type": "string"
																},
																"customerName": {
																	"type": "string"
																},
																"orderDate": {
																	"type": "string",
																	"format": "date-time"
																},
																"id": {
																	"type": "number",
																	"format": "double"
																}
															},
															"required": [
																"status",
																"customerAddress",
																"customerPhone",
																"customerEmail",
																"customerName",
																"orderDate",
																"id"
															],
															"type": "object"
														}
													]
												},
												"success": {
													"type": "boolean"
												}
											},
											"required": [
												"newOrder",
												"success"
											],
											"type": "object"
										},
										{
											"properties": {
												"newOrder": {},
												"status": {
													"type": "number",
													"format": "double"
												},
												"error": {
													"type": "string"
												},
												"success": {
													"type": "boolean"
												}
											},
											"required": [
												"status",
												"error",
												"success"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Orders"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Order"
							}
						}
					}
				}
			}
		},
		"/order/get": {
			"get": {
				"operationId": "CheckOrder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"error": {},
												"findOrder": {},
												"status": {
													"type": "number",
													"format": "double"
												},
												"success": {
													"type": "boolean"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"status",
												"success",
												"message"
											],
											"type": "object"
										},
										{
											"properties": {
												"error": {},
												"status": {},
												"message": {},
												"findOrder": {
													"items": {
														"allOf": [
															{
																"properties": {
																	"orderItems": {
																		"items": {
																			"properties": {
																				"orderId": {
																					"type": "number",
																					"format": "double"
																				},
																				"productId": {
																					"type": "number",
																					"format": "double"
																				},
																				"quantity": {
																					"type": "number",
																					"format": "double"
																				},
																				"id": {
																					"type": "number",
																					"format": "double"
																				}
																			},
																			"required": [
																				"orderId",
																				"productId",
																				"quantity",
																				"id"
																			],
																			"type": "object"
																		},
																		"type": "array"
																	}
																},
																"required": [
																	"orderItems"
																],
																"type": "object"
															},
															{
																"properties": {
																	"status": {
																		"type": "string"
																	},
																	"customerAddress": {
																		"type": "string"
																	},
																	"customerPhone": {
																		"type": "string"
																	},
																	"customerEmail": {
																		"type": "string"
																	},
																	"customerName": {
																		"type": "string"
																	},
																	"orderDate": {
																		"type": "string",
																		"format": "date-time"
																	},
																	"id": {
																		"type": "number",
																		"format": "double"
																	}
																},
																"required": [
																	"status",
																	"customerAddress",
																	"customerPhone",
																	"customerEmail",
																	"customerName",
																	"orderDate",
																	"id"
																],
																"type": "object"
															}
														]
													},
													"type": "array"
												},
												"success": {
													"type": "boolean"
												}
											},
											"required": [
												"findOrder",
												"success"
											],
											"type": "object"
										},
										{
											"properties": {
												"findOrder": {},
												"message": {},
												"status": {
													"type": "number",
													"format": "double"
												},
												"error": {
													"type": "string"
												},
												"success": {
													"type": "boolean"
												}
											},
											"required": [
												"status",
												"error",
												"success"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Orders"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "customerName",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "customerEmail",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "customerPhone",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "customerAdress",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/order/getAll": {
			"get": {
				"operationId": "GetAllOrder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"status": {},
												"allOrder": {},
												"totalOrder": {},
												"success": {
													"type": "boolean"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"success",
												"message"
											],
											"type": "object"
										},
										{
											"properties": {
												"message": {},
												"status": {
													"type": "number",
													"format": "double"
												},
												"allOrder": {
													"items": {
														"allOf": [
															{
																"properties": {
																	"orderItems": {
																		"items": {
																			"properties": {
																				"orderId": {
																					"type": "number",
																					"format": "double"
																				},
																				"productId": {
																					"type": "number",
																					"format": "double"
																				},
																				"quantity": {
																					"type": "number",
																					"format": "double"
																				},
																				"id": {
																					"type": "number",
																					"format": "double"
																				}
																			},
																			"required": [
																				"orderId",
																				"productId",
																				"quantity",
																				"id"
																			],
																			"type": "object"
																		},
																		"type": "array"
																	}
																},
																"required": [
																	"orderItems"
																],
																"type": "object"
															},
															{
																"properties": {
																	"status": {
																		"type": "string"
																	},
																	"customerAddress": {
																		"type": "string"
																	},
																	"customerPhone": {
																		"type": "string"
																	},
																	"customerEmail": {
																		"type": "string"
																	},
																	"customerName": {
																		"type": "string"
																	},
																	"orderDate": {
																		"type": "string",
																		"format": "date-time"
																	},
																	"id": {
																		"type": "number",
																		"format": "double"
																	}
																},
																"required": [
																	"status",
																	"customerAddress",
																	"customerPhone",
																	"customerEmail",
																	"customerName",
																	"orderDate",
																	"id"
																],
																"type": "object"
															}
														]
													},
													"type": "array"
												},
												"totalOrder": {
													"type": "number",
													"format": "double"
												},
												"success": {
													"type": "boolean"
												}
											},
											"required": [
												"status",
												"allOrder",
												"totalOrder",
												"success"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Orders"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}